[REFACTOR] 공용 컴포넌트 푸터를 직접 작성으로 변경
