+++ src/layouts/mobile-layout.tsx
@@ -1,6 +1,9 @@
-import { Outlet, useLocation } from "react-router-dom";
-import Footer from "../components/common/Footer";
+import { Outlet, useLocation, useNavigate } from "react-router-dom";
+import Background from "../assets/icons/footer-background.svg";
+import Home from "../assets/icons/home.svg";
+import Note from "../assets/icons/note-edit.svg";
 import UpButton from "../components/common/UpButton";
+import CalendarIcon from "../components/common/icon/Calendar";
 
 export default function MobileLayout() {
   const location = useLocation();
@@ -8,11 +11,54 @@ export default function MobileLayout() {
 
   const showFooter = footerPaths.includes(location.pathname);
 
+  const navigate = useNavigate();
+
   return (
     <div className="flex min-h-[100dvh] justify-center bg-gray-50">
       <div className="relative w-full max-w-[400px] bg-white shadow-md">
         <Outlet />
-        {showFooter && <Footer />}
+        {showFooter && (
+          <div className="fixed bottom-0 left-1/2 -translate-x-1/2 w-full max-w-[400px] text-[10px] font-medium z-10 text-center">
+            <div className="relative">
+              <img src={Background} className="w-full" alt="푸터 배경" />
+
+              <div className="absolute bottom-0 w-full aspect-[400/92] flex justify-between px-6">
+                <div className="flex justify-center w-19 pt-2">
+                  <button
+                    className="flex flex-col cursor-pointer gap-2"
+                    onClick={() => navigate("/")}
+                  >
+                    <img src={Home} className="size-5" alt="홈 아이콘" />
+                    <p>홈</p>
+                  </button>
+                </div>
+
+                <div className="flex flex-col items-center w-19 pt-2">
+                  <button
+                    className="flex flex-col items-center cursor-pointer gap-2"
+                    onClick={() => navigate("/meal-plan")}
+                  >
+                    <CalendarIcon size={20} />
+                    <p>식단표</p>
+                  </button>
+                </div>
+              </div>
+
+              <button
+                className="
+            absolute left-1/2 -translate-x-1/2 bottom-[clamp(32px,calc(56/400*100vw),56px)]
+            flex flex-col justify-center items-center gap-1 
+            w-[clamp(40px,calc(68/400*100vw),56px)]
+            h-[clamp(40px,calc(68/400*100vw),56px)] 
+            rounded-full bg-primary-700 cursor-pointer"
+                onClick={() => navigate("/meal-plan/create")}
+              >
+                <img src={Note} className="size-5" alt="식단짜기 아이콘" />
+                <p className="text-white">식단 짜기</p>
+              </button>
+            </div>
+          </div>
+        )}
         <UpButton />
       </div>
     </div>
